property income and expenses

+++ propertyManagement/src/model/relation/ProperMonthRecord.java
@@ -4,40 +4,40 @@ package model.relation;
  * Created by your dad on 2019/1/3.
  */
 public class ProperMonthRecord {
-    private int year;
-    private int month;
-    private int type;
-    private float amount;
+    private Integer year;
+    private Integer month;
+    private Integer type;
+    private Float amount;
 
-    public int getYear() {
+    public Integer getYear() {
         return year;
     }
 
-    public void setYear(int year) {
+    public void setYear(Integer year) {
         this.year = year;
     }
 
-    public int getMonth() {
+    public Integer getMonth() {
         return month;
     }
 
-    public void setMonth(int month) {
+    public void setMonth(Integer month) {
         this.month = month;
     }
 
-    public int getType() {
+    public Integer getType() {
         return type;
     }
 
-    public void setType(int type) {
+    public void setType(Integer type) {
         this.type = type;
     }
 
-    public float getAmount() {
+    public Float getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(Float amount) {
         this.amount = amount;
     }
 }
